ZCS-2344:ZMailbox split at mountpoint
`getParentFolderStoreAndUnmatchedPart` attempts to split a path into the
folder of the containing mountpoint and the sub-folder path remaining.
This was not happening and just a folder representing a folder in
another user's mailbox is returned.

Note comment in previous code talked about using a more efficient
algorithm but not convinced that was true (even though it was my
comment)